+++ common/src/main/java/net/frozenblock/wilderwild/registry/WWBlockModifications.java
@@ -26,7 +26,6 @@ import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.RotatedPillarBlock;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 
-// TODO: NeoForge
 public final class WWBlockModifications implements BlockRegistryModificationEvents.ModifyProperties, BlockRegistryModificationEvents.ReplaceFactory {
 
 	@Override